Optimizing metric value writes
Rather than sorting the whole list when we add a new metric value, we iterate through the existing list, copying the values to a new list, and inserting the new metric value when we reach the position that it’s sort order dictates that is should go. This changes the write time complexity from a worst case of On^2 (depending on the sorting algorithm) to On.